littletools_speech/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "littletools-speech"
version = "0.1.0"
description = "Speech processing tools for LittleTools project."
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "littletools-core",
    "setuptools>=65.0.0",
    "wheel>=0.38.0",
    "torch>=2.0.0",
    "torchaudio>=2.0.0",
    "openai-whisper @ git+https://github.com/openai/whisper.git",
    "ffmpeg-python>=0.2.0",
    "typer[all]>=0.9.0",
    "tqdm",
]

[project.scripts]
whisper-transcriber = "littletools_speech.whisper_transcriber:app"

[tool.setuptools.packages.find]
where = ["."]

[project.entry-points."littletools.commands"]
whisper-transcriber = "littletools_speech.whisper_transcriber:app" 